+++ shortest-path-problem/MinArray.h
@@ -8,6 +8,7 @@
  */
 class MinArray : public PriorityQueue
 {
+
 private:
     Pair* m_Data;
     bool* m_Flags;
@@ -20,9 +21,9 @@ public:
     MinArray();
     ~MinArray();
     
-    void Build(int* i_InitArray, int i_NumOfVertices) override;
+    void Build(float* i_InitArray, int i_NumOfVertices) override;
     bool IsEmpty() const override { return m_QueueSize == 0; }
-    void DecreaseKey(int i_Vertex, int i_NewKey) override;
+    void DecreaseKey(int i_Vertex, float i_NewKey) override;
     int DeleteMin() override;
     std::string GetPriorityQueueName() const override;
 };
